PodShare/requirements.txt
Flask==2.3.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
